Removed duplication of ExtBufWrapper class
