+++ samples/sample_common/include/sample_utils.h
@@ -27,12 +27,16 @@ or https://software.intel.com/en-us/media-client-solutions-support.
 #include <map>
 #include <stdexcept>
 #include <mutex>
+#include <algorithm>
 
 #include "mfxstructures.h"
 #include "mfxvideo.h"
 #include "mfxvideo++.h"
 #include "mfxjpeg.h"
 #include "mfxplugin.h"
+#include "mfxbrc.h"
+#include "mfxfei.h"
+#include "mfxfeihevc.h"
 
 #include "vm/strings_defs.h"
 #include "vm/file_defs.h"
@@ -588,7 +592,23 @@ mfxStatus DARtoPAR(mfxU32 darw, mfxU32 darh, mfxU32 w, mfxU32 h, mfxU16 *pparw, 
 // function for getting a pointer to a specific external buffer from the array
 mfxExtBuffer* GetExtBuffer(mfxExtBuffer** ebuffers, mfxU32 nbuffers, mfxU32 BufferId);
 
-//declare used extended buffers
+// class is used as custom exception
+class mfxError : public std::runtime_error
+{
+public:
+    mfxError(mfxStatus status = MFX_ERR_UNKNOWN, std::string msg = "")
+        : runtime_error(msg)
+        , m_Status(status)
+    {}
+
+    mfxStatus GetStatus() const
+    { return m_Status; }
+
+private:
+    mfxStatus m_Status;
+};
+
+//declare used extension buffers
 template<class T>
 struct mfx_ext_buffer_id{
     enum {id = 0};
@@ -611,7 +631,45 @@ template<>struct mfx_ext_buffer_id<mfxExtAVCRefListCtrl>{
 template<>struct mfx_ext_buffer_id<mfxExtThreadsParam>{
     enum {id = MFX_EXTBUFF_THREADS_PARAM};
 };
-
+template<>struct mfx_ext_buffer_id<mfxExtFeiParam> {
+    enum { id = MFX_EXTBUFF_FEI_PARAM };
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiPreEncCtrl> {
+    enum { id = MFX_EXTBUFF_FEI_PREENC_CTRL };
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiPreEncMV>{
+    enum {id = MFX_EXTBUFF_FEI_PREENC_MV};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiPreEncMBStat>{
+    enum {id = MFX_EXTBUFF_FEI_PREENC_MB};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncFrameCtrl>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_CTRL};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncMVPredictors>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_MV_PRED};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncQP>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_QP};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncCtuCtrl>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_CTU_CTRL};
+};
+template<>struct mfx_ext_buffer_id<mfxExtHEVCRefLists>{
+    enum {id = MFX_EXTBUFF_HEVC_REFLISTS};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcRepackCtrl>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_REPACK_CTRL};
+};
+template<>struct mfx_ext_buffer_id<mfxExtFeiHevcRepackStat>{
+    enum {id = MFX_EXTBUFF_HEVCFEI_REPACK_STAT};
+};
+template<>struct mfx_ext_buffer_id<mfxExtBRC> {
+    enum {id = MFX_EXTBUFF_BRC};
+};
+template<>struct mfx_ext_buffer_id<mfxExtHEVCParam> {
+    enum {id = MFX_EXTBUFF_HEVC_PARAM};
+};
 
 //helper function to initialize mfx ext buffer structure
 template <class T>
@@ -621,6 +679,198 @@ void init_ext_buffer(T & ext_buffer)
     reinterpret_cast<mfxExtBuffer*>(&ext_buffer)->BufferId = mfx_ext_buffer_id<T>::id;
     reinterpret_cast<mfxExtBuffer*>(&ext_buffer)->BufferSz = sizeof(ext_buffer);
 }
+
+/** ExtBufHolder is an utility class which
+ *  provide interface for mfxExtBuffer objects management in any mfx structure (e.g. mfxVideoParam)
+ */
+template<typename T>
+class ExtBufHolder : public T
+{
+public:
+    ExtBufHolder() : T()
+    {
+    }
+
+    ~ExtBufHolder() // only buffers allocated by wrapper can be released
+    {
+        for (auto it = m_ext_buf.begin(); it != m_ext_buf.end(); it++ )
+        {
+            delete [] (mfxU8*)(*it);
+        }
+    }
+
+    ExtBufHolder(const ExtBufHolder& ref)
+    {
+        *this = ref; // call to operator=
+    }
+
+    ExtBufHolder& operator=(const ExtBufHolder& ref)
+    {
+        const T* src_base = &ref;
+        return operator=(*src_base);
+    }
+
+    ExtBufHolder(const T& ref)
+    {
+        *this = ref; // call to operator=
+    }
+
+    ExtBufHolder& operator=(const T& ref)
+    {
+        // copy content of main structure type T
+        T* dst_base = this;
+        const T* src_base = &ref;
+        Copy(*dst_base, *src_base);
+
+        //remove all existing extension buffers
+        ClearBuffers();
+
+        //reproduce list of extension buffers and copy its content
+        m_ext_buf.reserve(ref.NumExtParam);
+        for (size_t i = 0; i < ref.NumExtParam; ++i)
+        {
+            const mfxExtBuffer* src_buf = ref.ExtParam[i];
+            if (!src_buf) throw mfxError(MFX_ERR_NULL_PTR, "Null pointer attached to source ExtParam");
+            if (!IsCopyAllowed(src_buf->BufferId)) throw mfxError(MFX_ERR_UNDEFINED_BEHAVIOR, "Copying buffer with pointers not allowed");
+
+            mfxExtBuffer* dst_buf = AddExtBuffer(src_buf->BufferId, src_buf->BufferSz);
+            if (!dst_buf) throw mfxError(MFX_ERR_UNDEFINED_BEHAVIOR, "Can't allocate destination buffer");
+            // copy buffer content w/o restoring its type
+            memcpy((void*)dst_buf, (void*)src_buf, src_buf->BufferSz);
+        }
+
+        return *this;
+    }
+
+    ExtBufHolder(ExtBufHolder &&)             = default;
+    ExtBufHolder & operator= (ExtBufHolder&&) = default;
+
+    template<typename TB>
+    TB* AddExtBuffer()
+    {
+        mfxExtBuffer* b = AddExtBuffer(mfx_ext_buffer_id<TB>::id, sizeof(TB));
+        return (TB*)b;
+    }
+
+    void RemoveExtBuffer(mfxU32 id)
+    {
+        auto it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
+        if (it != m_ext_buf.end())
+        {
+            delete [] (mfxU8*)(*it);
+            m_ext_buf.erase(it);
+            RefreshBuffers();
+        }
+    }
+
+    template <typename TB>
+    TB* GetExtBuffer()
+    {
+        auto it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(mfx_ext_buffer_id<TB>::id));
+        return it != m_ext_buf.end() ? (TB*)*it : nullptr;
+    }
+
+    template <typename TB>
+    operator TB*()
+    {
+        return (TB*)GetExtBuffer(mfx_ext_buffer_id<TB>::id);
+    }
+
+    template <typename TB>
+    operator TB*() const
+    {
+        return (TB*)GetExtBuffer(mfx_ext_buffer_id<TB>::id);
+    }
+
+private:
+
+    mfxExtBuffer* AddExtBuffer(mfxU32 id, mfxU32 size)
+    {
+        if (!size || !id)
+            return nullptr;
+
+        // Limitation: only one ExtBuffer instance can be stored
+        auto it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
+        if (it == m_ext_buf.end())
+        {
+            mfxExtBuffer* buf = (mfxExtBuffer*)new mfxU8[size];
+            memset(buf, 0, size);
+            m_ext_buf.push_back(buf);
+
+            mfxExtBuffer& ext_buf = *buf;
+            ext_buf.BufferId = id;
+            ext_buf.BufferSz = size;
+            RefreshBuffers();
+
+            return m_ext_buf.back();
+        }
+
+        return *it;
+    }
+
+    mfxExtBuffer* GetExtBuffer(mfxU32 id) const
+    {
+        const auto it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
+        return it != m_ext_buf.end() ? *it : nullptr;
+    }
+
+    void RefreshBuffers()
+    {
+        this->NumExtParam = (mfxU32)m_ext_buf.size();
+        this->ExtParam    = this->NumExtParam ? m_ext_buf.data() : nullptr;
+    }
+
+    void ClearBuffers()
+    {
+        if (m_ext_buf.size())
+        {
+            for (auto it = m_ext_buf.begin(); it != m_ext_buf.end(); it++ )
+            {
+                delete [] (mfxU8*)(*it);
+            }
+            m_ext_buf.clear();
+        }
+        RefreshBuffers();
+    }
+
+    bool IsCopyAllowed(mfxU32 id)
+    {
+        static const mfxU32 allowed[] = {
+            MFX_EXTBUFF_CODING_OPTION,
+            MFX_EXTBUFF_CODING_OPTION2,
+            MFX_EXTBUFF_CODING_OPTION3,
+            MFX_EXTBUFF_FEI_PARAM,
+            MFX_EXTBUFF_BRC,
+            MFX_EXTBUFF_HEVC_PARAM,
+        };
+
+        auto it = std::find_if(std::begin(allowed), std::end(allowed),
+                               [&id](const mfxU32 allowed_id)
+                               {
+                                   return allowed_id == id;
+                               });
+        return it != std::end(allowed);
+    }
+
+    struct CmpExtBufById
+    {
+        mfxU32 id;
+
+        CmpExtBufById(mfxU32 _id)
+            : id(_id)
+        { };
+
+        bool operator () (mfxExtBuffer* b)
+        {
+            return  (b && b->BufferId == id);
+        };
+    };
+
+    std::vector<mfxExtBuffer*> m_ext_buf;
+};
+
+typedef ExtBufHolder<mfxVideoParam> MfxVideoParamsWrapper;
+typedef ExtBufHolder<mfxEncodeCtrl> mfxEncodeCtrlWrap;
 
 // returns false if buf length is insufficient, otherwise
 // skips step bytes in buf with specified length and returns true
@@ -781,35 +1031,6 @@ mfxI32 getMonitorType(msdk_char* str);
 void WaitForDeviceToBecomeFree(MFXVideoSession& session, mfxSyncPoint& syncPoint, mfxStatus& currentStatus);
 
 mfxU16 FourCCToChroma(mfxU32 fourCC);
-
-// class is used as custom exception
-class mfxError
-{
-public:
-    mfxError(mfxStatus status = MFX_ERR_UNKNOWN, std::string msg = "")
-        : m_Status(status),
-          m_msg(msg)
-    {
-    }
-
-    virtual ~mfxError()
-    {
-    }
-
-    mfxStatus GetStatus() const
-    {
-        return m_Status;
-    }
-
-    std::string GetMessage() const
-    {
-        return m_msg;
-    }
-
-private:
-    mfxStatus m_Status;
-    std::string m_msg;
-};
 
 #ifdef ENABLE_MCTF
 // this function implements a simple management of MCTF control-buffers that can be attached to pmfxSurface
