+++ samples/sample_hevc_fei_abr/include/sample_hevc_fei_defs.h
@@ -197,225 +197,6 @@ template<class T> inline T Clip3(T min, T max, T x) { return std::min(std::max(m
 
 /**********************************************************************************/
 
-template<>struct mfx_ext_buffer_id<mfxExtFeiParam> {
-    enum { id = MFX_EXTBUFF_FEI_PARAM };
-};
-template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncFrameCtrl>{
-    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_CTRL};
-};
-template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncMVPredictors>{
-    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_MV_PRED};
-};
-template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncQP>{
-    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_QP};
-};
-template<>struct mfx_ext_buffer_id<mfxExtFeiHevcEncCtuCtrl>{
-    enum {id = MFX_EXTBUFF_HEVCFEI_ENC_CTU_CTRL};
-};
-template<>struct mfx_ext_buffer_id<mfxExtHEVCRefLists>{
-    enum {id = MFX_EXTBUFF_HEVC_REFLISTS};
-};
-
-struct CmpExtBufById
-{
-    mfxU32 id;
-
-    CmpExtBufById(mfxU32 _id)
-        : id(_id)
-    { };
-
-    bool operator () (mfxExtBuffer* b)
-    {
-        return  (b && b->BufferId == id);
-    };
-};
-
-/** ExtBufWrapper is an utility class which
- *  provide interface for mfxExtBuffer objects management in any mfx structure (e.g. mfxVideoParam)
- */
-template<typename T>
-class ExtBufWrapper: public T
-{
-public:
-    ExtBufWrapper()
-    {
-        T& base = *this;
-        Zero(base);
-    }
-
-    ~ExtBufWrapper() // only buffers allocated by wrapper can be released
-    {
-        for(ExtBufIterator it = m_ext_buf.begin(); it != m_ext_buf.end(); it++ )
-        {
-            delete [] (mfxU8*)(*it);
-        }
-    }
-
-    ExtBufWrapper(const ExtBufWrapper& ref)
-    {
-        *this = ref; // call to operator=
-    }
-
-    ExtBufWrapper& operator=(const ExtBufWrapper& ref)
-    {
-        const T* src_base = &ref;
-        return operator=(*src_base);
-    }
-
-    explicit ExtBufWrapper(const T& ref)
-    {
-        *this = ref; // call to operator=
-    }
-
-    ExtBufWrapper& operator=(const T& ref)
-    {
-        // copy content of main structure type T
-        T* dst_base = this;
-        const T* src_base = &ref;
-        Copy(*dst_base, *src_base);
-
-        //remove all existing extended buffers
-        ClearBuffers();
-
-        //reproduce list of extended buffers and copy its content
-        m_ext_buf.reserve(ref.NumExtParam);
-        for (size_t i = 0; i < ref.NumExtParam; ++i)
-        {
-            const mfxExtBuffer* src_buf = ref.ExtParam[i];
-            if (!src_buf) throw mfxError(MFX_ERR_NULL_PTR, "Null pointer attached to source ExtParam");
-            if (!IsCopyAllowed(src_buf->BufferId)) throw mfxError(MFX_ERR_UNDEFINED_BEHAVIOR, "Copying buffer with pointers not allowed");
-
-            mfxExtBuffer* dst_buf = AddExtBuffer(src_buf->BufferId, src_buf->BufferSz);
-            // copy buffer content w/o restoring its type
-            memcpy((void*)dst_buf, (void*)src_buf, src_buf->BufferSz);
-        }
-
-        return *this;
-    }
-
-    template<typename TB>
-    TB* AddExtBuffer()
-    {
-        mfxExtBuffer* b = AddExtBuffer(mfx_ext_buffer_id<TB>::id, sizeof(TB));
-        return (TB*)b;
-    }
-
-    void RemoveExtBuffer(mfxU32 id)
-    {
-        ExtBufIterator it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
-        if (it != m_ext_buf.end())
-        {
-            delete [] (mfxU8*)(*it);
-            m_ext_buf.erase(it);
-            RefreshBuffers();
-        }
-    }
-
-    template <typename TB>
-    TB* GetExtBuffer()
-    {
-        ExtBufIterator it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(mfx_ext_buffer_id<TB>::id));
-        if (it != m_ext_buf.end())
-        {
-            return (TB*)*it;
-        }
-        return 0;
-    }
-
-    template <typename TB>
-    operator TB*()
-    {
-        return (TB*)GetExtBuffer(mfx_ext_buffer_id<TB>::id);
-    }
-
-    template <typename TB>
-    operator TB*() const
-    {
-        return (TB*)GetExtBuffer(mfx_ext_buffer_id<TB>::id);
-    }
-
-private:
-    mfxExtBuffer* AddExtBuffer(mfxU32 id, mfxU32 size)
-    {
-        if (!size || !id)
-            throw mfxError(MFX_ERR_UNDEFINED_BEHAVIOR, "Failed to attach undefined buffer");
-
-        // Limitation: only one ExtBuffer instance can be stored
-        ExtBufIterator it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
-        if (it == m_ext_buf.end())
-        {
-            mfxExtBuffer* buf = (mfxExtBuffer*)new mfxU8[size];
-            memset(buf, 0, size);
-            m_ext_buf.push_back(buf);
-
-            mfxExtBuffer& ext_buf = *buf;
-            ext_buf.BufferId = id;
-            ext_buf.BufferSz = size;
-            RefreshBuffers();
-
-            return m_ext_buf.back();
-        }
-
-        return *it;
-    }
-
-    mfxExtBuffer* GetExtBuffer(mfxU32 id) const
-    {
-        constExtBufIterator it = std::find_if(m_ext_buf.begin(), m_ext_buf.end(), CmpExtBufById(id));
-        if (it != m_ext_buf.end())
-        {
-            return *it;
-        }
-        return 0;
-    }
-
-    void RefreshBuffers()
-    {
-        this->NumExtParam = (mfxU32)m_ext_buf.size();
-        this->ExtParam    = this->NumExtParam ? m_ext_buf.data() : 0;
-    }
-
-    void ClearBuffers()
-    {
-        if (0 != m_ext_buf.size())
-        {
-            for (ExtBufIterator it = m_ext_buf.begin(); it != m_ext_buf.end(); it++ )
-            {
-                delete [] (mfxU8*)(*it);
-            }
-            m_ext_buf.clear();
-        }
-        RefreshBuffers();
-    }
-
-    bool IsCopyAllowed(mfxU32 buf_id)
-    {
-        static mfxU32 allowed[] = {
-            MFX_EXTBUFF_CODING_OPTION,
-            MFX_EXTBUFF_CODING_OPTION2,
-            MFX_EXTBUFF_CODING_OPTION3,
-            MFX_EXTBUFF_FEI_PARAM,
-        };
-
-        for (mfxU32 i = 0; i < sizeof(allowed)/sizeof(allowed[0]); i++)
-        {
-            if (buf_id == allowed[i])
-                return true;
-        }
-
-        return false;
-    }
-
-    typedef std::vector<mfxExtBuffer*>::iterator ExtBufIterator;
-    typedef std::vector<mfxExtBuffer*>::const_iterator constExtBufIterator;
-    std::vector<mfxExtBuffer*> m_ext_buf;
-};
-
-typedef ExtBufWrapper<mfxVideoParam> MfxVideoParamsWrapper;
-typedef ExtBufWrapper<mfxEncodeCtrl> mfxEncodeCtrlWrap;
-
-/**********************************************************************************/
-
 inline mfxU32 align(const mfxU32 val, const mfxU32 alignment)
 {
     STATIC_ASSERT(!(val == 0) && !(val & (val - 1)), is_power_of_2);
